refactor: Unify field and variant attributes
