+++ crates/stackable-versioned-macros/src/attrs/variant.rs
@@ -1,8 +1,8 @@
 use convert_case::{Case, Casing};
 use darling::{Error, FromVariant};
-use syn::{Ident, Variant};
+use syn::Ident;
 
-use crate::attrs::common::{ContainerAttributes, ItemAttributes, ItemType};
+use crate::attrs::common::{ItemAttributes, ItemType};
 
 #[derive(Debug, FromVariant)]
 #[darling(
@@ -52,58 +52,5 @@ impl VariantAttributes {
 
         errors.finish()?;
         Ok(self)
-    }
-
-    pub(crate) fn validate_versions(
-        &self,
-        container_attrs: &ContainerAttributes,
-        variant: &Variant,
-    ) -> Result<(), Error> {
-        // NOTE (@Techassi): Can we maybe optimize this a little?
-        // TODO (@Techassi): Unify this with the field impl, e.g. by introducing
-        // a T: Spanned bound for the second function parameter.
-        let mut errors = Error::accumulator();
-
-        if let Some(added) = &self.common.added {
-            if !container_attrs
-                .versions
-                .iter()
-                .any(|v| v.name == *added.since)
-            {
-                errors.push(Error::custom(
-                   "variant action `added` uses version which was not declared via #[versioned(version)]")
-                   .with_span(&variant.ident)
-               );
-            }
-        }
-
-        for rename in &*self.common.renames {
-            if !container_attrs
-                .versions
-                .iter()
-                .any(|v| v.name == *rename.since)
-            {
-                errors.push(
-                   Error::custom("variant action `renamed` uses version which was not declared via #[versioned(version)]")
-                   .with_span(&variant.ident)
-               );
-            }
-        }
-
-        if let Some(deprecated) = &self.common.deprecated {
-            if !container_attrs
-                .versions
-                .iter()
-                .any(|v| v.name == *deprecated.since)
-            {
-                errors.push(Error::custom(
-                   "variant action `deprecated` uses version which was not declared via #[versioned(version)]")
-                   .with_span(&variant.ident)
-               );
-            }
-        }
-
-        errors.finish()?;
-        Ok(())
     }
 }
